feat: toast para erros no crud e exportação de usuário

+++ app/Http/Services/User/UserService.php
@@ -4,6 +4,7 @@ namespace App\Http\Services\User;
 
 use App\Http\Repositories\User\UserRepositoryInterface;
 use App\Models\User;
+use Exception;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Facades\Hash;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
@@ -29,6 +30,12 @@ class UserService implements UserServiceInterface
 
     public function store(array $data): User
     {
+        $userWithIdentificationAlreadyExists = $this->userRepository->getByIdentification($data['identification']);
+
+        if ($userWithIdentificationAlreadyExists) {
+            throw new Exception('Já existe um usuário com essa identificação.');
+        }
+
         $data['password'] = Hash::make('123');
 
         return $this->userRepository->store($data);
@@ -36,11 +43,29 @@ class UserService implements UserServiceInterface
 
     public function update(int $id, array $data): bool
     {
+        $userWithIdentificationAlreadyExists = $this->userRepository->getByIdentification($data['identification']);
+
+        if ($userWithIdentificationAlreadyExists && $userWithIdentificationAlreadyExists->id !== $id) {
+            throw new Exception('Já existe um usuário com essa identificação.');
+        }
+
         return $this->userRepository->update($id, $data);
     }
 
     public function destroy(int $id): bool
     {
+        $user = $this->userRepository->getById($id);
+
+        $hasMetrologyCalls = false;
+
+        if ($user->userRole->name === 'Metrologista' || $user->userRole->name === 'Operador') {
+            $hasMetrologyCalls = $user->openedMetrologyCalls()->count() > 0 || $user->closedMetrologyCalls()->count() > 0;
+        }
+
+        if ($user->logs()->count() > 0 || $hasMetrologyCalls) {
+            throw new Exception('Não é possível excluir um usuário que possui registros de logs ou chamados de metrologia.');
+        }
+        
         return $this->userRepository->destroy($id);
     }
